pyproject.toml
[tool.poetry]
name = "redisenv"
version = "0.1.0"
description = ""
authors = ["Redis OSS <oss@redis.io>"]
readme = "README.md"
license = "MIT"

include = [
    {path = "redisenv/templates/*", format=["sdist", "wheel"]}
]

classifiers = [
    "Programing Language :: Python",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
repository = "https://github.com/RedisLabsModules/redisenv"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
click = "^8.1.3"
docker = "^6.0.1"
loguru = "^0.6.0"
semver = "^2.13.0"
jinja2 = "^3.1.2"
pyyaml = "^6.0"

[tool.poetry.scripts]
redisenv = "redisenv:main"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
invoke = "^2.0.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
